Switch parsing function
The *arc* module, in some combination of recent versions of SWIG, Python
and ARC, does not have the *arc.JobDescription_Parse* function anymore.
It is not understood why, but it was discovered
that *arc.JobDescription.Parse* is and equivalent function. It appears
as if SWIG is not reexporting static functions into top-level module
anymore, maybe because reexportation was not meant to happen.